Add support for robots that produce multiple files
Some robots like /video/adaptive and /document/thumbs produce multiple
files. Previously we weren't handling this, since Transloadit returns an
array for both single and multiple files, we would always just save
the first processed file.

With this change it's not possible to specify that the result of a
step will be multiple files via the #multiple declaration:

  class DocumentUploader < Shrine
    plugin :versions

    def transloadit_process(io, context)
      thumbs = transloadit_file(io)
        .add_step("thumbs", "/document/thumbs")
        .multiple

      { thumbs: thumbs }
    end
  end

Now the result will be saved as multiple indexed versions:

  document.file[:thumbs_0]
  document.file[:thumbs_1]
  ...
  document.file[:thumbs_n]

Closes #2